refactor asce7-16 and nehrp-2015 into nshmp-2014 factory
