+++ test/nehrp-2015-factory.spec.js
@@ -4,34 +4,7 @@
 
 const NEHRP2015Factory = require('../src/lib/nehrp-2015-factory'),
     expect = require('chai').expect;
-    //sinon = require('sinon');
 
-const _DUMMY_FACTORY = {
-  metadataFactory: {
-    getMetadata: () => { return Promise.resolve([]); }
-  },
-  probabilisticService: {
-    getData: () => { return Promise.resolve({}); }
-  },
-  deterministicService: {
-    getData: () => { return Promise.resolve({}); }
-  },
-  riskCoefficientService: {
-    getData: () => { return Promise.resolve({}); }
-  },
-  tSubLService: {
-    getData: () => { return Promise.resolve({response: {data: {}}}); }
-  },
-  siteAmplificationFactory: {
-    getSiteAmplificationData: () => { return Promise.resolve([]); }
-  },
-  designCategoryFactory: {
-    getDesignCategory: () => { return Promise.resolve([]); }
-  },
-  spectraFactory: {
-    getSpectrum: () => { return Promise.resolve([]); }
-  }
-};
 
 describe('nehrp-2015-factory', () => {
   describe('constructor', () => {
@@ -54,23 +27,6 @@ describe('nehrp-2015-factory', () => {
 
       const factory = NEHRP2015Factory();
       expect(factory.referenceDocument).to.equal('NEHRP-2015');
-    });
-  });
-
-  describe('get', () => {
-    it('returns a promise and calls functions as intended', () => {
-
-      const factory = NEHRP2015Factory(_DUMMY_FACTORY);
-
-      let result = factory.get({
-        latitude: 0,
-        longitude: 1,
-        riskCategory: 2,
-        siteClass: 'C',
-        title: 'Test'
-      });
-
-      expect(result).to.be.instanceof(Promise);
     });
   });
 });
